+++ move_ave.py
